feat: also separate MCC split case endpoint
